Create application
I designed the user interface and declared a textField outlet as well as actions for the start and stop buttons in the AppDelegate.rb.

My current problem is that I cannot assign the textField outlet to my text field nor the actions to any of my buttons through the Interface Builder.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>